requirements.txt
torch==2.0.1
torchvision
streamlit
matplotlib
Pillow
